+++ src/index.ts
@@ -10,6 +10,8 @@ import {Message} from "./client/models";
 import {TemplatedMessage} from "./client/models";
 import {Attachment} from "./client/models";
 import {Header} from "./client/models";
+import {BulkEmailRequest} from "./client/models";
+import {BulkEmailMessage} from "./client/models";
 
 export {ServerClient, ServerClient as Client, AccountClient, AccountClient as AdminClient, Models, Errors};
-export {Message, TemplatedMessage, Attachment, Header};
+export {Message, TemplatedMessage, Attachment, Header, BulkEmailRequest, BulkEmailMessage};